pyproject.toml
[tool.poetry]
name = "discord-reminderbot"
version = "0.1.0"
description = "Record reminders from people in Nightshift discord"
authors = ["Tyler Stubenvoll <tstubenv@fastmail.fm>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
xdice = "^1.2.1"

[tool.poetry.dev-dependencies]
wheel = "^0.36.2"
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
